Moved NFTDetails and VHRToken to VHR Table Renamed VHRToken to be more consistent
