+++ src/components/Profile/VHRTable/NFTDetails.tsx
